requirements.txt
absl-py==2.1.0
accelerate==0.26.1
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backcall==0.2.0
bleach==6.1.0
blis==0.7.11
braceexpand==0.1.7
catalogue==2.0.10
certifi==2023.11.17
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
cmake==3.25.0
confection==0.1.4
contextlib2==21.6.0
contexttimer==0.3.3
contourpy==1.1.1
cycler==0.12.1
cymem==2.0.8
datasets==2.16.1
decorator==5.1.1
decord==0.6.0
deepspeed==0.12.2
dill==0.3.7
distlib==0.3.8
docker-pycreds==0.4.0
einops==0.7.0
executing==2.0.1
fairscale==0.4.4
filelock==3.13.1
flash-attn @ file:///mnt/vepfs/fs_users/lisihang/3D-MoLM/flash_attn-2.3.1.post1%2Bcu117torch2.0cxx11abiFALSE-cp38-cp38-linux_x86_64.whl#sha256=783a0f6f73494feda956e0233023dad92b43856603902494b6acd161820af13f
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2023.10.0
ftfy==6.1.3
gitdb==4.0.11
GitPython==3.1.41
hjson==3.1.0
huggingface-hub==0.20.3
identify==2.5.33
idna==3.6
importlib-resources==6.1.1
iopath==0.1.10
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
kaggle==1.6.3
kiwisolver==1.4.5
langcodes==3.3.0
lightning-utilities==0.10.1
lit==15.0.7
littleutils==0.2.2
lmdb==1.4.1
MarkupSafe==2.1.3
matplotlib==3.7.4
matplotlib-inline==0.1.6
ml-collections==0.1.1
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.10
networkx==3.0
ninja==1.11.1.1
nltk==3.8.1
nodeenv==1.8.0
numpy==1.24.4
ogb==1.3.6
omegaconf==2.3.0
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
outdated==0.2.2
packaging==23.2
pandas==2.0.3
parso==0.8.3
peft==0.5.0
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
platformdirs==4.1.0
plotly==5.18.0
portalocker==2.8.2
pre-commit==3.5.0
preshed==3.0.9
prompt-toolkit==3.0.43
protobuf==4.25.2
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==15.0.0
pyarrow-hotfix==0.6
pycocoevalcap==1.2
pycocotools==2.0.7
pydantic==2.5.3
pydantic_core==2.14.6
pyg-lib==0.3.1+pt20cu117
Pygments==2.17.2
pynvml==11.5.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-slugify==8.0.1
pytorch-lightning==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
rdkit==2023.3.3
regex==2023.12.25
requests==2.31.0
rouge-score==0.1.2
safetensors==0.4.1
salesforce-lavis==1.0.2
scikit-learn==1.3.2
scipy==1.10.1
sentencepiece==0.1.99
sentry-sdk==1.39.2
setproctitle==1.3.3
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
sympy==1.12
tenacity==8.2.3
tensorboardX==2.6.2.2
text-unidecode==1.3
thinc==8.2.2
threadpoolctl==3.2.0
timm==0.9.12
tokenizers==0.15.0
torch==2.0.1+cu117
torch-cluster==1.6.3+pt20cu117
torch-scatter==2.1.2+pt20cu117
torch-sparse==0.6.18+pt20cu117
torch-spline-conv==1.2.2+pt20cu117
torch_geometric==2.3.1
torchmetrics==1.3.0.post0
torchvision==0.15.2+cu117
tqdm==4.66.1
traitlets==5.14.1
transformers==4.36.2
triton==2.0.0
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
unicore @ file:///mnt/vepfs/fs_users/lisihang/3D-MoLM/unicore-0.0.1%2Bcu117torch2.0.0-cp38-cp38-linux_x86_64.whl#sha256=6ab80bf8b04f90f66298ee41a2333c14c414e7ecdb98bc55575e66f8934d6a08
urllib3==2.1.0
virtualenv==20.25.0
wandb==0.16.2
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
webdataset==0.2.86
webencodings==0.5.1
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
